--- services/baseUrl.js
@@ -1,5 +1,0 @@
-const baseUrl = "http://localhost:8902/garagem-api/cadastro-garagem/";
-
-const baseUrl2 = "http://localhost:8901/pagamentos-api/formas-de-pagamento/";
-
-export { baseUrl, baseUrl2 };
